island: Move config manipulator functions from `config.py` to `config_manipulator.py`
